app/requirements.txt
fastapi
mysql-connector-python
email-validator
uvicorn
dotenv